target/rockchip: hinlink opc h68k/h69k fix gmac1 phy initialize failed
Since gmac may be initialized earlier than pmic, vccio1 is not powered at this time.
If the reset gpio of phy is on vccio1, phy reset cannot be triggered,
resulting in phy initialization failure.

So we should ensure that the corresponding IO domain is powered on before using gpio,
but there is currently no power supply configuration for gpio.
This workaround adds a phy-supply that forces the power supply of vccio1 to be relied on, so that gmac can be started after vccio1 is powered on.

If there is a better solution in the future, roll back this change.

Logs found on h68k:
```
...
[    1.237018] mdio_bus stmmac-1: MDIO device at address 0 is missing.
...
[    1.573974] rk808-regulator rk808-regulator.2.auto: DMA mask not set
[    1.590149] vdda0v9_image: Bringing 600000uV into 950000-950000uV
[    1.601310] vccio_acodec: Bringing 600000uV into 3300000-3300000uV
[    1.619220] vcca1v8_image: Bringing 600000uV into 1800000-1800000uV
...
[   12.749856] rk_gmac-dwmac fe010000.ethernet eth0: Register MEM_TYPE_PAGE_POOL RxQ-0
[   12.751008] rk_gmac-dwmac fe010000.ethernet eth0: __stmmac_open: Cannot attach to PHY (error: -19)
```